Fixed e621 command (FINALLY) and managed to made it dynamic instead of Hardcoded, for real this time

+++ sbot.py
@@ -8,7 +8,7 @@ import configparser
 #this is most definitely not the way, but this is how Imma do it for now till I figure things out
 logging.basicConfig(level=logging.INFO)
 
-if not os.path.isfile("config.ini") and not os.path.isfile("e621useragent.txt"):
+if not os.path.isfile("config.ini"):
     botconfig = configparser.ConfigParser()
 
     botconfig["Bot Config"] = {
@@ -20,26 +20,29 @@ if not os.path.isfile("config.ini") and not os.path.isfile("e621useragent.txt"):
 
     with open('config.ini', 'w') as conf:
         botconfig.write(conf)
-    open('e621useragent.txt', 'w')
+
+if not os.path.isfile("e621useragent.txt"):
+    useragent = open('e621useragent.txt', 'w')
+    useragent.write('#e621 user agent format: {"User-Agent": "Your Agent Name /Your Agent Version (By Your Name)"}')
     print("Config files created, please read and edit very carefully to make sure that the bot works")
     exit()
-else:
-     print(f"Process is running with PID = {os.getpid()}")
 
-     bot = commands.AutoShardedBot
+print(f"Process is running with PID = {os.getpid()}")
 
-     bot = bot(command_prefix=getconfig('prefix'),
-     help_command=None
+bot = commands.AutoShardedBot
+
+bot = bot(command_prefix=getconfig('prefix'),
+help_command=None
 )
 
-     for x in commandregister:
-         bot.load_extension(x)
-     for x in eventregister:
-         bot.load_extension(x)
+for x in commandregister:
+    bot.load_extension(x)
+for x in eventregister:
+    bot.load_extension(x)
 
-     bot.run(
-        getconfig('token')
+bot.run(
+    getconfig('token')
 )
-     bot.owner_id(
-        getconfig("ownerid")
+bot.owner_id(
+    getconfig("ownerid")
 )
